+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - job list command
 - deployment list command
 - service list command
+- events command
 - version command
 
 ## [0.7.0] - 2023-06-26